pyproject.toml
[project]
name = "faster"
version = "0.1.0"
description = "RPN training on xView dataset using Faster R-CNN"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "rasterio>=1.3.0",
    "shapely>=2.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.10.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
